auth-service/pyproject.toml
[tool.poetry]
name = "auth-service"
version = "0.1.0"
description = ""
authors = ["Muhammad-Hasham-Khalid <m.hasham681@gmail.com>"]
readme = "README.md"
packages = [{include = "auth_service"}]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.92.0"
uvicorn = "^0.20.0"
pyjwt = "^2.6.0"
python-decouple = "^3.7"
sqlalchemy = "^2.0.4"
pydantic = {extras = ["email"], version = "^1.10.5"}
psycopg2-binary = "^2.9.5"
fastapi-sqlalchemy = "^0.2.1"
bcrypt = "^4.0.1"


[tool.poetry.group.dev.dependencies]
black = "^23.1.0"

[tool.poetry.scripts]
dev = "auth_service.main:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
